requirements.txt
fastapi[all];
pillow;
uvicorn;
cssmin;
jsmin;
requests;